requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==2.6.1
numpy==1.26.4
pdfplumber==0.11.4
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0
openai==1.35.13
anthropic==0.71.0
chromadb==1.2.1
routellm==0.1.0
pdf2image==1.16.3
langchain==0.0.330
pypdf==3.17.2
gunicorn==21.2.0
psutil==5.9.8
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict